Additional layered fees
#7
This reworks the message passing in the Arena protocol.
Instead of passing Arena tax config around, let's just store this in the Arena core and return when querying for tax. From here, we can send an ordered list of layered fees to automatically send when distributing an escrow. Additional layered fees are saved directly on the competition and are optional.
Also cleans up states on migrations.